몬스터 추적

+++ Oculus/Assets/02_Scirpts/Enemy/EnemyBase.cs
@@ -6,8 +6,9 @@ public class EnemyBase : MonoBehaviour
 {
     public Player player;
 
+
     [Header("Stat")]
-    public int enemy_Hp;
+    public float enemy_Hp;
     public float enemy_MoveSpeed;
-    public int enemy_Damage;
+    public float enemy_Damage;
 }
